+++ newLSTMpreprocess.py
